backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
mistralai>=1.0
faiss-cpu>=1.7
pymupdf>=1.23
fpdf2>=2.7
numpy>=1.26
tenacity>=8.0
orjson>=3.9
httpx>=0.27
